Update hints
I hope everyone is doing well with the exercise so far. If you see this commit message, I would like to let you know that there is something hidden in this Markdown document if you haven't found it yet.

Here is the hint, "Markdown also has a comment." Good luck, have fun!